Don't send anything as retained
Reading the documentation on the algorithm about how to design a Victron
UI, MQTT section, we'd have to re-add stuff about ignoring messages with
the retain flag. This is something that proved to be too error prone,
so it's better to just not use retain.

At some point, we could decide to always publish certain paths on a
timer though.