Fix new paths appearing in a service through PropertiesChanged
Behavior that was seen was that when settings paths appear, by changing
the relay function to 'pump' for instance, it would throw an error that
it didn't recognize those paths as MQTT items, because they hadn't been
seen in initial scans.

The path for PropertiesChanged is now the same as ItemsChanged, so it
behaves the same, and items are properly created.

It was also confirmed now that removing settings with the dbus call
'RemoveSettings', unpublishes the MQTT topics, like:

N/<portalid>/settings/0/Settings/Pump0/AutoStartEnabled {"value":null}
N/<portalid>/settings/0/Settings/Pump0/Mode {"value":null}
N/<portalid>/settings/0/Settings/Pump0/StartValue {"value":null}
N/<portalid>/settings/0/Settings/Pump0/StopValue {"value":null}
N/<portalid>/settings/0/Settings/Pump0/TankService {"value":null}

Note that's it's not an empty MQTT publish. This is a result of an empty
array of the proper int type being published on the dbus. An empty MQTT
publish does happen on service disappearance, but there is no equivalent
of path disappearance.