Fix dealing with signals that don't contain the values of paths
Apparently there are situations where we get signals that only contains
'Mix', 'Max', or 'Default' (as individual signals). This fix prevents
taking the uninitialized value in those cases and assigning to an item.

This fixes values with 'unknown_type'.

One occurance this was seen specifically, was for settings there were
changed by the user. Upon boot of the device these signals would be
fired.