Exit the main process on dbus out-of-memory errors
It was observed that because of the DBusPendingCall leak (fixed a few
commits back), that once this situation exists, logs are spammed and it
doesn't recover.

On Venus, care is taken that dbus is always available, or restarted. Our
own limits are lower because we run under softlimit, so we just need
restart.