Disconnect users on signal UserRemoved from Venus Platform
For that, we need to track the users that log in with their clientids
(for lookup later), and purge the list periodically.

This change makes sure we immediately disconnect an MQTT connection when
a paired device is unpaired (when the token is removed).

Example MQTT use of this on Platform:

* Topic: W/<portalid>/platform/0/Tokens/Remove
* Payload: { "value": "token/evcharger/mydevice123" }